requirements.txt
absl-py==0.8.0
appnope==0.1.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bert-serving-client==1.9.8
bert-serving-server==1.9.8
boto==2.49.0
boto3==1.9.235
botocore==1.12.235
bpemb==0.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
Deprecated==1.2.6
docutils==0.15.2
elasticsearch==7.0.4
Flask==1.1.1
future==0.17.1
gast==0.3.2
gensim==3.8.0
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.23.0
h5py==2.10.0
hyperopt==0.1.2
idna==2.8
importlib-metadata==0.23
ipython==7.6.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
langdetect==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
mpld3==0.3
networkx==2.3
numpy==1.17.2
packaging==19.2
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.9
protobuf==3.9.2
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.1.3
python-dateutil==2.8.0
pytorch-transformers==1.2.0
pytz==2019.2
pyzmq==18.1.0
regex==2019.8.19
requests==2.22.0
s3transfer==0.2.1
sacremoses==0.0.34
scikit-learn==0.21.3
scipy==1.3.1
segtok==1.5.7
sentencepiece==0.1.83
six==1.12.0
sklearn==0.0
smart-open==1.8.4
sqlitedict==1.6.0
tabulate==0.8.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.2.0
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
